Add --force-confnew to apt-get install
When apt has a problem installing a package because a config file
has changed, it will ask the user what to do. This requires stdin
which breaks the downstream container build.

I think (though this is a bit of a guess) that we'll always want the
new, packaged version of the config file.